requirements.txt
# 现实预警星座多智能体滚动任务规划系统 v2.0.0 依赖包

# Google ADK框架
google-adk>=0.1.0

# 核心依赖
asyncio-mqtt>=0.11.0
aiohttp>=3.8.0
websockets>=10.0

# LLM集成
litellm>=1.0.0
openai>=1.0.0

# 数据处理
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0

# 可视化
matplotlib>=3.5.0
plotly>=5.0.0
seaborn>=0.11.0
kaleido>=0.2.1

# 配置管理
PyYAML>=6.0
python-dotenv>=0.19.0

# 时间处理
python-dateutil>=2.8.0

# Web框架 (UI)
flask>=2.0.0
flask-socketio>=5.0.0

# STK集成
comtypes>=1.1.0
pywin32>=227

# 日志处理
colorlog>=6.0.0

# 数学计算
sympy>=1.9.0

# 网络通信
requests>=2.25.0
urllib3>=1.26.0

# 并发处理
concurrent-futures>=3.1.0

# 系统监控
psutil>=5.8.0
memory-profiler>=0.60.0

# 测试框架
pytest>=6.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0

# 开发工具
black>=21.0.0
flake8>=4.0.0
mypy>=0.910

# 文档生成
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0

# 数据序列化
pickle-mixin>=1.0.0
dill>=0.3.0

# 加密和安全
cryptography>=3.4.0

# 性能优化
cython>=0.29.0
numba>=0.56.0

# 图形界面 (可选)
tkinter-tooltip>=2.0.0

# 数据库 (可选)
sqlite3

# 其他工具
tqdm>=4.62.0
rich>=10.0.0
